requirements.txt
natto-py==0.9.0
sentencepiece==0.1.82
torch==1.1.0